+++ docassemble_base/docassemble/base/error.py
@@ -7,7 +7,10 @@ class DAError(Exception):
 
 class CodeExecute(Exception):
     def __init__(self, compute, question):
-        self.compute = compute
+        if type(compute) is list:
+            self.compute = "\n".join(compute)
+        else:
+            self.compute = compute
         self.question = question
     
 class ForcedNameError(NameError):